Refactored Compeleted
