+++ src/api/profile/index.ts
@@ -1,11 +1,6 @@
+import { getRequest } from "api/axios"
 import { ProfileRoutes } from "api/profile/routes"
-import { AxiosResponse } from "axios"
-import axios from "api/axios"
 
-export async function setUser(): Promise<AxiosResponse> {
-	let response: AxiosResponse
-	response = await axios.get(ProfileRoutes.get_user)
-	localStorage.removeItem("user")
-	localStorage.setItem("user", JSON.stringify(response.data))
-	return response
+export async function initUser() {
+	return getRequest(ProfileRoutes.get_user)
 }
