+++ src/api/profile/routes.ts
@@ -1,3 +1,3 @@
 export enum ProfileRoutes {
-	get_user = "auth/get-user/",
+	get_user = "/auth/get-user/",
 }
