+++ src/components/Player/style.ts
