--- src/layouts/Player/emotion.ts
@@ -1,8 +1,0 @@
-import { styled } from "@mui/material"
-
-export const Container = styled("div")`
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  max-width: 80em;
-`
