+++ src/pages/PlayerScreen/emotion.ts
@@ -1,0 +1,9 @@
+import styled from "@emotion/styled/macro"
+
+export const Container = styled("div")`
+  display: flex;
+  align-items: flex-start;
+  max-width: 80vw;
+  flex-direction: column;
+  gap: 4em;
+`
