+++ src/pages/PlayerScreen/index.tsx
@@ -1,14 +1,18 @@
 import React, { FC } from "react"
-import Player from "layouts/Player"
+import Player from "components/Player"
 import { useParams } from "react-router-dom"
+import VideoGrid from "components/VideoGrid"
+import { Container } from "./emotion"
 
 const PlayerScreen: FC = () => {
 	const { videoId } = useParams()
 
 	return (
-		<>
+		<Container>
 			<Player videoId={videoId as string} />
-		</>
+
+			<VideoGrid />
+		</Container>
 	)
 }
