+++ src/utils/index.ts
@@ -13,3 +13,4 @@ export { logOut } from "utils/user"
 export { setRefreshToken } from "utils/user"
 export { getAccessToken } from "utils/user"
 export { setAccessToken } from "utils/user"
+export { setUser } from "utils/user"
